Isolate pre-read helper stack for safer adapter flow
Move the pre-read envelope, payload-plan, debug, and RN/WebView boundary helper stack into a sibling adapter module so decidePreRead stays focused on runtime orchestration while preserving existing output contracts.

Constraint: This cut must be behavior-preserving and must not move domain or payload policy semantics out of their current core owners.

Rejected: Extract runtime adapters or domain policies in the same PR | that would widen the shared-file blast radius beyond the helper-stack seam.

Confidence: high

Scope-risk: narrow

Directive: Keep decidePreRead as the orchestration boundary; future helper movement should preserve boundary fallback ordering before payload planning.

Tested: npm run build; npm run typecheck -- --pretty false; focused pre-read/fooks/runtime/payload-policy tests; npm test; git diff --check; support-claim grep over docs and src.

Not-tested: Live provider runtime outside local harness tests.
