+++ test/pre-read-fallback-builder.test.mjs
@@ -11,6 +11,7 @@ const repoRoot = process.cwd();
 const require = createRequire(import.meta.url);
 const preRead = require(path.join(repoRoot, "dist", "adapters", "pre-read.js"));
 const preReadSource = fs.readFileSync(path.join(repoRoot, "src", "adapters", "pre-read.ts"), "utf8");
+const preReadStackSource = fs.readFileSync(path.join(repoRoot, "src", "adapters", "pre-read-stack.ts"), "utf8");
 
 function assertFallbackOnlyDecision(decision, reason) {
   assert.equal(decision.decision, "fallback");
@@ -23,8 +24,9 @@ function assertFallbackOnlyDecision(decision, reason) {
 }
 
 test("pre-read centralizes full-read fallback envelope construction", () => {
-  assert.match(preReadSource, /function buildPreReadFallbackDecision\(/);
-  const fallbackEnvelopeConstructions = preReadSource.match(/fallback:\s*{\s*\n\s*action:\s*"full-read"/g) ?? [];
+  assert.match(preReadStackSource, /function buildPreReadFallbackDecision\(/);
+  assert.match(preReadSource, /buildPreReadFallbackDecision/);
+  const fallbackEnvelopeConstructions = preReadStackSource.match(/fallback:\s*{\s*\n\s*action:\s*"full-read"/g) ?? [];
   assert.equal(fallbackEnvelopeConstructions.length, 1);
   assert.doesNotMatch(preReadSource, /return \{\s*\n\s*runtime,[\s\S]*?decision: "fallback",[\s\S]*?fallback: \{\s*\n\s*action: "full-read"/);
 });
@@ -47,14 +49,14 @@ test("pre-read fallback builder preserves ineligible extension decisions", () =>
 });
 
 test("pre-read source-shape boundary guard skips payload planning", () => {
-  assert.match(preReadSource, /function hasWebViewSourceShapeBoundary\(/);
-  assert.match(preReadSource, /function shouldUseReactNativeWebViewBoundaryFallback\(/);
+  assert.match(preReadStackSource, /function hasWebViewSourceShapeBoundary\(/);
+  assert.match(preReadStackSource, /function shouldUseReactNativeWebViewBoundaryFallback\(/);
   assert.match(preReadSource, /if \(shouldUseReactNativeWebViewBoundaryFallback\(domainDetection\)\) \{/);
-  const sourceShapeGuardIndex = preReadSource.indexOf("function shouldUseReactNativeWebViewBoundaryFallback(");
+  const sourceShapeGuardIndex = preReadStackSource.indexOf("function shouldUseReactNativeWebViewBoundaryFallback(");
   const boundaryGuardIndex = preReadSource.indexOf("if (shouldUseReactNativeWebViewBoundaryFallback(domainDetection))");
   const payloadPlanIndex = preReadSource.indexOf('const { payload, readiness, debug } = buildPreReadPayloadPlan({');
   assert.ok(sourceShapeGuardIndex >= 0);
-  assert.ok(boundaryGuardIndex > sourceShapeGuardIndex);
+  assert.ok(boundaryGuardIndex >= 0);
   assert.ok(payloadPlanIndex > boundaryGuardIndex);
 
   const tempDir = fs.mkdtempSync(path.join(repoRoot, ".tmp-pre-read-source-shape-"));
